refactor(type): 将 ClassConstructor 类型定义统一迁移至 AirType 文件中
- 从多个文件中移除重复的 ClassConstructor 类型定义
- 在 AirType.ts 文件中添加统一的 ClassConstructor 类型定义
- 更新相关文件中的导入路径，使用新的 ClassConstructor 类型
