+++ config/AirConfig.ts
@@ -9,8 +9,7 @@ import { INormalTreeProps } from '../interface/props/INormalTreeProps'
 import { IUser } from '../interface/IUser'
 import { AirFileEntity } from '../model/entity/AirFileEntity'
 import { AirUserEntity } from '../model/entity/AirUserEntity'
-import { ClassConstructor } from '../type/ClassConstructor'
-import { AirCodeNumber, AirMoneyDirection } from '../type/AirType'
+import { AirCodeNumber, AirMoneyDirection, ClassConstructor } from '../type/AirType'
 import { AirApi } from './AirApi'
 import { AirConstant } from './AirConstant'
