+++ interface/decorators/ISearchFieldConfig.ts
@@ -3,8 +3,7 @@ import { IField } from './IField'
 import { AirSearchDataType } from '../../enum/AirSearchDataType'
 import { AirDictionaryArray } from '../../model/extend/AirDictionaryArray'
 import { AirEnum } from '../../base/AirEnum'
-import { ClassConstructor } from '../../type/ClassConstructor'
-import { AirEnumKey } from '../../type/AirType'
+import { AirEnumKey, ClassConstructor } from '../../type/AirType'
 
 /**
  * # 表格的字段配置接口
