+++ type/AirType.ts
@@ -133,3 +133,12 @@ export type AirIconType =
   | 'MONITOR'
   | 'COMPARE'
   | 'CLOSE'
+
+/**
+ * ## 类包装
+ * @author Hamm.cn
+ */
+export type ClassConstructor<T = AirAny> = {
+  // eslint-disable-next-line no-unused-vars
+  new(...args: AirAny[]): T;
+}
